readFitness() method in Evolver now takes a file, reads it, and creates the FitnessList. To test this, I created a inputData.txt file containing the data that is currently being used as our sample data, then pasted the location of the file directly into the call from GPMain to Evolve.readFitness(). I believe what we will actually want in the long run is arg[0], so this is what is currently in place, but if I am wrong, but all means correct me.
Also, it should be noted that in order to read a file correctly, the
data in the file must have a very specific format. This format is
demonstrated by the following:
x:0 y:0 0
x:0 y:1 1
x:1 y:0 1
And so forth. If this is not how we want to accept data, obviously so
things will need to be changed, but the framework is there at least.

Finally, while FitnessList appears to be correct, computeFitness()
method in Fitness is now breaking, as it claims the inputList is null
and therefore does not have a size. I spent a while trying to figure out
why this was, and came up with nothing, although I imagine it has to be
some issue with the FitnessList generated by readFitness(), because it
works fine with the data hardcoded in. I will continue to try to figure
this issue out.

All tests pass.