Changed all the "id" labels to "uid" labels. I had thought this would avoid a problem where our "id" property was shadowing "id" as the name for the numeric node number that Neo4J assigns, but I'm not entirely sure that was necessary.
I also cast the fitness coming from the child to a double before setting
it as a Neo4J node property. This was necessary to get some of the tests
to pass as from the tests (but never from the production code?!?) we
were returning a BigDemical as the fitness and Neo4J was throwing an
exception when we tried to set it as a property value.

Lastly, I replaced all the tabs with spaces, which reformatted the whole
darn thing, which was kind of annoying because then it was really hard
to see what "real" things I'd changed.