Made Individual cloneable using @AutoClone.
Changed a bug in setUid() where we weren't actually setting a new UID
for that Individual; we really should have that happen automatically
when we clone an Individual instead of requiring a call to setUid().

Removed compareTo() since (a) I think it was broken (it should return an
numeric value and it return an individual) and (b) we never actually
called it.